Remove casting into an int (#750)

+++ lib/debezium/types_test.go
@@ -132,7 +132,7 @@ func TestField_ParseValue(t *testing.T) {
 				Type: Int32,
 			},
 			value:         float64(3),
-			expectedValue: 3,
+			expectedValue: int64(3),
 		},
 		{
 			name: "decimal",
